src/test/java/com/epam/finaltack/ivanvertylo/web/command/AdminTest.java
package com.epam.finaltack.ivanvertylo.web.command;

import com.epam.finaltack.ivanvertylo.Path;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

import static org.junit.Assert.assertEquals;

@RunWith(MockitoJUnitRunner.class)
public class AdminTest {

    private Admin admin;
    @Mock
    private HttpServletRequest request;
    @Mock
    private HttpServletResponse response;

    @Before
    public void before(){
        admin = new Admin();
    }

    @Test
    public void testExecute() throws IOException, ServletException {
        String result = admin.execute(request,response);
        assertEquals(result, Path.CONTROLLER_ADMIN_PAGE);
    }

}
